% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_form}
\alias{create_form}
\title{Create A Google Form}
\usage{
create_form(json_file, title, token)
}
\arguments{
\item{json_file}{The json_file}

\item{title}{The title of the form}
}
\value{
resp
}
\description{
Create A Google Form
}
