% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_question}
\alias{create_question}
\title{Create Form Questions}
\usage{
create_question(json_file, form_id, items)
}
\arguments{
\item{json_file}{The json file}

\item{form_id}{The id of the form}

\item{items}{The list of questions}
}
\value{
question_resp
}
\description{
Create Form Questions
}
