% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_hash_key}
\alias{generate_hash_key}
\title{Generate Hash Key}
\usage{
generate_hash_key(num_of_forms, orig_link)
}
\arguments{
\item{num_of_forms}{The number of Google forms}
}
\value{
dat
}
\description{
Generate Hash Key
}
