% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_response}
\alias{process_response}
\title{Process A Response To Extract Form ID}
\usage{
process_response(resp)
}
\arguments{
\item{resp}{The response from the Google Forms API request.}
}
\value{
form_id
}
\description{
This extract the form ID from the response.
}
