% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_form}
\alias{update_form}
\title{Update Form Description}
\usage{
update_form(json_file, form_id)
}
\arguments{
\item{json_file}{The json_file}

\item{form_id}{The id of the form}

\item{items}{The list of items}
}
\value{
form_update_resp
}
\description{
Update Form Description
}
